requirements-test.txt
# CI/CDテスト用依存関係
# requirements.txtから重いパッケージ（ta-lib, freqtrade）のみ除外

# Core Framework
fastapi==0.109.1
uvicorn[standard]==0.25.0
gunicorn==21.2.0
python-multipart==0.0.6

# Auth & Security
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
cryptography>=42.0.5
age==0.5.1

# Config & Utilities
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
PyYAML==6.0.1
orjson==3.9.10
click==8.1.7
cachetools==5.3.2

# Database & Supabase
sqlalchemy==2.0.25
supabase==2.4.2
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Data Processing
pandas==2.1.4
numpy==1.26.3
pyarrow==14.0.2
polars==0.20.2

# Trading & Market Data
ccxt==4.2.25

# Technical Analysis (Phase 3)
pandas-ta==0.3.14b
talib-binary==0.4.19

# freqtrade==2024.1  # CI除外：重すぎる
# ta==0.11.0        # ta-libの軽量版は含める
# ta-lib==0.4.28    # CI除外：コンパイルが重い

# Async & Networking
httpx==0.26.0
aiohttp==3.9.1
websockets==12.0
tenacity==8.2.3

# Monitoring & Logging
prometheus-client==0.19.0
rich==13.7.0
loguru==0.7.2

# Background Jobs
rq==1.15.1
redis==5.0.1
schedule==1.2.0
